Added MixVideoDifferentPlanes
There was the necesity to mix sagittal, frontal, and back view of
patient video recording. This mix videos are not synch since each
film was made with different cameras: iphone 4, samsung galaxy s3,
and Le3s. It needs to be improved sooner than expected with diffe-
rent and better color cameras.