Added markerRecognitionImprovement
The current method of maker detection has limitations which generate
issues with the 3D reconstruction algorithm. A novel method was im-
plemented using a simple math calculation with RGB format to segment
the pixels of a retroreflectivea marker. The effects of this
change are positive: i) a light way of segmentation process, ii) no
need of thresh calculation, and iii) better result can be achieved
if the brightness of retroreflective markers are higher.