Created Json that holds kinematic angles to Highcharts object
The 3D algorithm did not generate kinematic data in Highchart format.
Methods were implemented which create a json to be send via API
to Gaitcome.com. The impact of this is the contribution with the
integration of Matlab with gaitcome server. Jsonencode and Jsonde-
code were used to acomplish the task.