Updated consume PUT method to send data through request header
It was not possible to consume PUT method because of internal Matlab
error:
	"ICU error: get canonical time zone ID,
	err='U_ILLEGAL_ARGUMENT_ERROR'.

So, the solution provided was to send the Body as a Request header
attribute to the server side using the Online Matlab version. The
impoct of this is the interoperability of Online Matlab and Gaitcome
server side.