Hot-fix Sync error calculated
Due to misunderstanding variable names,the current sync method is
not accurate. It was necessary code refactory to have a better
understading of the algorithm. During this process, the approach
slighly changed: the threshold was decreased to a value of 200.
The effect of this change provides better results in the bliking
comparison process throwing from 95% to 98% of sync frames.