Added Absolute angles
There were some issues when calculation the kinematic angles:
inconsistencies in the angles vatiariation. Now a better angle
calculation was implemented: absolute angles which, basically,
calculte the angle in respect to a horizontal line. The effect
of this implementation is consistency in calculation the angles.
Furthermore, this change was test recalculation the saggital
angles: hip, plv, ank, and kne, repectively.