Test - using CLEYE sync images
Tracking system had not been tested with sync PsEye footage. The
images were retrived from Gaitcome2D window app. Small changes
were added to process the images in mention. The results of this
test are favorable for to reason. For one thing better calibration
was achieved thanks to the third sync software. Second, the point
cloud oscilation is less conspicuous than before. However, the
maker segmentation must be improved, for it is causing the point
cloud oscilation.